fix bf16 error
To solved the error "Got unsupported ScalarType BFloat16" when computing reward with bf16 dtype

+++ ImageReward/ImageReward.py
@@ -133,7 +133,7 @@ class ImageReward(nn.Module):
         rewards = self.mlp(txt_features)
         rewards = (rewards - self.mean) / self.std
         
-        return rewards.detach().cpu().numpy().item()
+        return rewards.detach().cpu().float().numpy().item()
 
 
     def inference_rank(self, prompt, generations_list):
